separate core and dropwizard bundle
