+++ qtrouper-core/src/main/java/io/github/qtrouper/Trouper.java
