+++ qtrouper-core/src/main/java/io/github/qtrouper/core/config/QueueConfiguration.java
