+++ qtrouper-core/src/main/java/io/github/qtrouper/core/config/SidelineConfiguration.java
