+++ qtrouper-core/src/main/java/io/github/qtrouper/core/models/QAccessInfo.java
