+++ qtrouper-core/src/main/java/io/github/qtrouper/core/rabbit/RabbitConfiguration.java
