+++ qtrouper-core/src/test/java/io/github/qtrouper/core/config/QueueConfigurationTest.java
