+++ qtrouper-dw/pom.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">  
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <artifactId>qtrouper</artifactId>
+    <groupId>io.github.qtrouper</groupId>
+    <version>0.0.2</version>
+  </parent>
+
+  <artifactId>qtrouper-dw</artifactId>
+  <packaging>jar</packaging>
+  <version>0.0.2</version>
+  <name>qtrouper-dropwizard-Bundle</name>
+
+  <properties>
+    <dropwizard.version>1.2.2</dropwizard.version>
+    <lombok.version>1.18.6</lombok.version>
+  </properties>
+
+
+  <dependencies>
+    <dependency>
+      <groupId>io.github.qtrouper</groupId>
+      <artifactId>qtrouper-core</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.dropwizard</groupId>
+      <artifactId>dropwizard-core</artifactId>
+      <version>${dropwizard.version}</version>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.dropwizard</groupId>
+      <artifactId>dropwizard-testing</artifactId>
+      <version>${dropwizard.version}</version>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
+
+</project>
